Revert "Move lots logic from my bids to saleCard"
This reverts commit 7a293decafffc99b6142188aea5ba6d0bb9500fe.
