+++ src/lib/Scenes/MyBids/Components/SaleCard.tsx
@@ -1,7 +1,7 @@
 import OpaqueImageView from "lib/Components/OpaqueImageView/OpaqueImageView"
 import { navigate } from "lib/navigation/navigate"
 import { SaleInfo } from "lib/Scenes/MyBids/Components/SaleInfo"
-import { ArrowRightIcon, Flex, Join, Separator, Text, Touchable } from "palette"
+import { ArrowRightIcon, Flex, Separator, Text, Touchable } from "palette"
 import React from "react"
 import { createFragmentContainer, graphql } from "react-relay"
 
@@ -9,9 +9,6 @@ import { ClockFill, ExclamationMarkCircleFill } from "palette/svgs/sf"
 
 import { SaleCard_me } from "__generated__/SaleCard_me.graphql"
 import { SaleCard_sale } from "__generated__/SaleCard_sale.graphql"
-
-import { ActiveLotFragmentContainer as ActiveLot, ClosedLotFragmentContainer as ClosedLot } from "."
-import { isLotStandingComplete } from "../helpers/timely"
 
 export const COVER_IMAGE_HEIGHT = 100
 
@@ -35,21 +32,14 @@ export const RegistrationCTAWrapper: React.FunctionComponent<{ navLink?: string 
   </Touchable>
 )
 
-const ContentWrapper: React.FC<{ smallScreen?: boolean }> = (props) => (
-  <>
-    <Separator mt={1} />
-    <Flex style={{ marginHorizontal: props.smallScreen! ? 10 : 20, marginVertical: 10 }}>{props.children}</Flex>
-  </>
-)
-
 export class SaleCard extends React.Component<{
   sale: SaleCard_sale
   me: SaleCard_me
   smallScreen?: boolean
-  lotStandings?: any[]
+  hideChildren?: boolean
 }> {
   render() {
-    const { sale, me, smallScreen, lotStandings } = this.props
+    const { sale, me, smallScreen, hideChildren, children } = this.props
     const { registrationStatus } = sale
 
     const pendingIdentityVerification = me.pendingIdentityVerification
@@ -96,24 +86,11 @@ export class SaleCard extends React.Component<{
               <SaleInfo sale={sale} />
               <RegistrationCTA />
             </Flex>
-            {!lotStandings && !!registrationStatus?.qualifiedForBidding && (
-              <ContentWrapper smallScreen={smallScreen}>
-                <Text color="black60" py={1} textAlign="center">
-                  You haven't placed any bids on this sale
-                </Text>
-              </ContentWrapper>
-            )}
-            {!!lotStandings && (
-              <ContentWrapper smallScreen={smallScreen}>
-                <Join separator={<Separator my={1} />}>
-                  {lotStandings.map((ls) => {
-                    if (ls && sale) {
-                      const LotInfoComponent = isLotStandingComplete(ls) ? ClosedLot : ActiveLot
-                      return <LotInfoComponent lotStanding={ls} key={ls?.lotState?.internalID} />
-                    }
-                  })}
-                </Join>
-              </ContentWrapper>
+            {!hideChildren && (
+              <>
+                <Separator mt={1} />
+                <Flex style={{ marginHorizontal: smallScreen! ? 10 : 20, marginVertical: 10 }}>{children}</Flex>
+              </>
             )}
           </Flex>
         </Touchable>
